Add grep facility while viewing the node and link tables. Scaling changes for introspect. Fix possibility of state being NULL.
For the node table, if the name of the node contains the search string, we
output it.
For the link table, if the names of the nodes of the link contain the search
string, we output it.
If no search string is given, we output everything.

Scaling changes:
1. Currently, the IFMapTableShowReq and IFMapTableShowReq introspects buffer
the entire table and then send in batches. This will not work in scaling
scenarios where the number of nodes/links is huge. Change both the introspects
to buffer/send 50 entries at a time and add code to iterate by providing a
next_batch link on the client page.
2. There are 3 cases in the introspect code where 'state' is dereferenced
without checking for NULL. Since, 'state' could be NULL when the introspect
catches the node while buffering, fix the code to access state only if its
non-NULL.
3. Add the table size for the link-table as part of the introspect output. The
link-table size output cannot be put in IFMapNodeTableListShowReq since the
table names in that show are clickable.
4. Add new APIs in the db-table/db-partition code needed by the above changes.

Change-Id: I51ddeb89fdf1374b34b4f782779b660b356e9368