Provide a way to bulk delete multiple links for an identifier
When an object is deleted, we currently walk all its references and delete each
of them one by one. Instead, we can delete all of them in one message to ifmap.
This code change provides a method to do it. In a subsequent change, we will add
a call to this method in auto-generated code.

Change-Id: I0dbf5c383cfef34388f97318536e873ea13a388d
Partial-Bug: 1432735