Align columns for bgp and xmpp neighbors in BgpNeighborListResp
Certain fields were made optional when ShowBgpNeighborSummaryReq command
was added.  Some of these fields were not populated for xmpp and others
were not populated for bgp.  This resulted in misaligned columns for bgp
and xmpp neighbors in BgpNeighborListResp.

Fix by explicitly setting such fields to empty value as appropriate i.e.
when it could be non-empty for one type of neighbor and empty for other.

Change-Id: I3ab8304856d4e3693c75616274b80c40e0f4bfad
Partial-Bug: #1445856