Send NodeStatusUve delete, when group is removed.
In cases where groups can be added & deleted, the NodeStatusUve also
needs to be added or deleted. In TSN case, TOR-Agents may be added
and deleted. Updated to set the delete status appropriately. The group
is deleted when the corresponding supervisor ini file is removed and
supervisorctl update is done.

Change-Id: I30b1ba38a635a7fc2417b5452f65cad9e657d207
Closes-bug: #1432951