Metadata route is not removed, causing VRF delete timeout.
When VN to VRF link is deleted and linklocal services are also deleted,
depending on the sequence, we may end up not deleting the routes added
for linklocal services. Adding the linklocal addresses in the state so
that it can be used to delete the linklocal routes.

Change-Id: Icfe51e64a44b299bdd89f286603d14fda9d7353c
Partial-bug: 1440990
(cherry picked from commit 9f0393db6ef7bcb54f286da236991de82524c785)