REST API for alarm acknowledgement in contrail-analytics-api
Added a new POST request for acknowledging alarms in analytics-api
http://<analytics-ip>:<rest-api-port>/analytics/alarms/acknowledge
1) json body for the POST request
 {
     “table”: <object-type>,
     “name”: <key>,
     “type”: <alarm type>,
     “token”: <token>
 }
"token" field is a base64 encoding of the following json.
{
    "host_ip" : <host-ip-address>,
    "http_port" : <introspect-port>,
    "timestamp" : <timestamp in UTC>
}

2) analytics-api upon receiving the alarm ack POST request would decode
the token to extract the generator endpoint and the timestamp.
It then sends the introspect request to the generator that raised this
alarm. Response to the client that sent the POST request is based on the
introspect response from the generator.
http://<generator-ip>:<introspect-port>/Snh_SandeshAlarmAckRequest?
table=<alarm-table>&name=<alarm-id>&type=<alarm-type>&timestamp=<timestamp>

Change-Id: I364d94bdf5cfe3a31471f9e35a59fd618afaca6f
Partial-Bug: #1447116