Use unique Id for the SNAT service instance
When stressing the snat with a loop setting/clearing
the GW the svc_monitor looses the references to the related
resources due to a race between the cleanup mechanism
of the svc and the service instance creation/deletion.

This patch generates an unique id for each service instance,
so that the svc is able to cleanup the related resource
properly. This unique id is stored in the KV. There is a
fallback to the old name in the clear method in order
to handle GW set before this patch.

Change-Id: I19502bfa77e4567946e9062be27f97e8766c5c84