Move SNAT code from VNC to SVC
This patch mainly moves SNAT code
related to the service instance
management from the VNC side to the
SVC side by introducing a snat
agent exaclty what has been done for
LBaaS.

This patch also introduces a base class
for all the agents and an agent manager
which handles and forward some evens during
the service instance creation. The goal
of this refactoring is to move the specific
pieces of code that are present in the
instance_manager closer to their related
agent.

For the snat part the SI fq_name is now
unique in order to limit race conditions
when setting/clearing GW.

WIP: waiting for feedback from the community
and UTs have to be introduced/fixed.

Change-Id: I19502bfa77e4567946e9062be27f97e8766c5c84