Fix TOR Agent Crash in KSync infra
Issue:
------
Following events happened in TOR Agent
Delete of an active unicast remote route entry, before
ack for this is received it triggered creation of stale
entry resulting in DB state stuck with ksync entry and
thus further DB operations on the same entry results in
a bad state.

Fix:
----
When a deleted KSyncEntry is converted to Stale Entry
remove its association with DB Entry, so that unless
KSyncEntry is reclaimed by removing stale entry flag
it should not trigger the state machine.

Added test-case for the same

Closes-Bug: 1451782
Change-Id: I1d4a86d5f73fbb882e6a34fe89b381d8c821aa4c