* Allocate vrf table label, only when there is a gateway interface
When there is a gateway interface in the VRF,
table label allocated for each VRF would be used while
exporting the route to control-node, upon hitting this
label routing would happen on the VRF where this label
points. Allocating the label only if there is a
gateway interface in that subnet.
Closes-bug:#1451744

Change-Id: I84d2f980a34af8affe9043731a7f4121faadcfe1