Child monitor daemon for vrouter_netns subprocess daemons
When a load balancer is created using vrouter_netns, it inturn spawns
a haproxy process which runs as daemon after the opencontrail_vrouter_netns exits.

But when the haproxy process is crashed for some reason, there is no way that
it will get respawned or make aware the agent about its crash. This patch
attempts to monitor the subprocess daemons that are created as part of the
opencontrail_vrouter_netns (now, its just haproxy).

Instead of daemonizing the subprocess, a low-weight child monitor process is
spawned which inturn spawns the haproxy. The child monitor process monitors
the daemons that are spawned using it and respawns the same in case of abnormal
exits of the subprocess daemons. When the suprocess daemons exits normally,
the child monitor too exits.

Closes-bug: #1424766
Change-Id: I9a4da703ab7f9824db4c010568bbb2d6ba994561