Fix auto-purge feature in analytics-api
Recently, database_usage in DatabaseUsageInfo structure was converted to
stats field to remove duplicate database_usage_stats field.
This change had resulted in exception while accessing the database_usage
as this field was converted to list and hence the auto-purge feature was broken.
Fixed the accessing of database_usage field.

Change-Id: I7e33a350e5428c8e3ccfd776b81e775bab99df8d
Closes-Bug: #1455021