Fix exporter code to handle delete-link followed by add-link.
We recently changed the code in IFMapLinkTable::AddLink() to fix 1426175 and
that fix is invalidating an assert in the exporter. In the case where we get a
delete-link followed by add-link, before the link is completely cleaned up, the
new code in IFMapLinkTable::AddLink() will find the older link and send a
change on it. This will result in the exporter asserting. Fix this by checking
if the link is being revived and calling appropriate asserts for the revival
case and the change case. Add test-case and fix test case that fails.

Change-Id: Ib594d73847c80b3e3d3f3650565b785d8aaf7f25
Closes-Bug: 1454380