DM UT code enhancements
DM code depends on the order of vnc api's executed, noticed some failures when order is changed.
Basic idea of this enhancement is divide the vnc-api execution blocks into multiple blocks,
and have a sequnce of blocks which needs to be executed. Python has the support for executing
code objects (code can be assigned to a variable, called code object). In our case, a sequence of
blocks can be considered as a UT case and we can have multiple sequences. After executing each sequnce,
we need to cleanup up vnc-db state, so that next Ut case sequence will execute without assuming any previous
state.

It works fine, its proposal.

Change-Id: Ic182b8391b4fc8d38ba1a94cadf8821197df3057