Make the API server report correct IP to discovery
The Contrail API will currently use the IP address that is
configured in contrail-api.conf as IP address to report to
the discovery service. The corresponding lines in
vnc_cfg_api_server.py are:

'ip-address': self._args.ifmap_server_ip,

and

'ip-address': self._args.ifmap_server_ip,

This is obviously a typo, as there is a separate parameter
available to specifiy the listen IP for the API, which is,
however, currently not used for discovery registration.

If the ifmap server IP is set to 127.0.0.1, the API will
report 127.0.0.1 as its listen-to address to discovery.

This patch introduces the correct behaviour and ensures that
the contrail API will correctly report its listen IP.

Change-Id: Iea4979443b94afaf7fd250854649b325ba48d4ef