DM: apply-groups should not delete any existing groups
The following netconf config statement removes all applied groups
and adds only '__contrail__' group. This is a problem.
<apply-groups operation="replace">__contrail__</apply-groups>

It is not required to send operation attribute, since there is only one
element which is a text '__contrail__' present always.
However, if it is for delete, then we must send 'delete' operation.

So, here is the logic:
if operation is not delete:
    <apply-groups>__contrail__</apply-groups>
else:
    <apply-groups operation="delete">__contrail__</apply-groups>

Closes-Bug: 1457265

Change-Id: Ida901177b2e68ae63c80b1c1433a3e9479aaa4c6