Fix access to bgp_peer after free
Bridge route for multicast MAC can contain multiple paths from
control-node.

When the connection from control-node breaks, we walk thru all route
tables and delete path from the peer. But, in DeletePathUsingKeyData we
delete only one path for a peer. This will result in a stale reference
to peer in the route.

Modified DeletePathUsingKeyData to delete all path from a peer.

Change-Id: Icd35349aa3ad3784351f884b533812e5c719e3ce
closes-bug:#1458057