Agent: XMPP connection not attempted by agent
Agent marks connection to Xmpp-Server DOWN after several attempts and rediscovers controllers.
If the controller list sent by Discovery is already marked DOWN, agent does not honor the
controller connection to be applied.  This happens in the time period where discovery has not
yet marked the controller down (marked when 3 heartbeats are missing) and sends the stale list.
Additionally as discovery client has a checksum so the callbacks are throttled at source and
hence the clients never get called when controllers are UP and the information is lost.

Solution is to honor reponse from discovery irrespective of the state of the publisher
(both Xmpp Server advertised by controller and dns daemon)

Test cases added.

Change-Id: I5c6695f04d3dd4a384f0ea7d0da912475966c6eb
Closes-Bug:1446463
Closes-Bug:1457243