Clearing service instance DB State after STOP command
If Delete of service instance is received before it was marked
unusable, StopServiceInstance command is issued, which invokes
the namespace script. And immediately after this, DB state is cleared.
Clearing of DB State can result in deleting of service instance
DBEntry before namespace script completion. SIGCHILD event after
script completion looks for service instance that triggered the
task, and results in crash as the service isntance is already deleted.
To avoid crash, the DBState is cleared after processing the SIGCHILD
event.

Testcases to follow in the next commit

Change-Id: I212063ede8f7df4b693771aa394c4536a76d8d20
closes-bug: #1453956