Clearing service instance DB State after STOP command
If Delete of service instance is received before it was marked
unusable, StopServiceInstance command is issued, which invokes
the namespace script. And immediately after this, DB state is cleared.
Clearing of DB State can result in deleting of service instance
DBEntry before namespace script completion. SIGCHILD event after
script completion looks for service instance that triggered the
task, and results in crash as the service isntance is already deleted.
To avoid crash, the DBState is cleared after processing the SIGCHILD
event.

Testcases to follow in the next commit

Change-Id: I406f9e3d70768151e4390955153bc3176d3da99f
closes-bug: 1449166, 1453956