Send interface list of Vrouter UVE only in master/primary TSN
In TSN, for each interface we check whether we have mastership for the corresponding physical router. If we have mastership, then that interface is send in interface_list of Vrouter UVE. Otherwise the interface is sent as part of new field unmanaged_list which is also part of Vrouter UVE. Also send VN count in Vrouter UVE.
Added and updated UT accordingly.

Change-Id: I033fadf0f72a1aebb0c229aab6103434870f38a3
Closes-Bug: #1459515