Allocate mac address for service instance ports
Service instance VM, netns do not request mac-addresses for its ports.
In cases of active-standby, active-active instances there is an issue
which can cause the standby mac-address to be picked up. This change
ensures that the same mac-address is used for all VMs using a particular
virtual network. This is achieved by allocating the mac-address extracted
from the virtual-network the port is attached to.

Change-Id: Ie96bd757f75f78b4aaa1c2a87a0e0ed2f902af01
Closes-Bug: #1460767