Allocate mac address for service instance ports
Service instance VM, netns do currently do not request mac-addresses
for its ports. In cases of active-standby, active-active instances there
is an issue which can cause the standby mac-address to be picked up because
of EVPN. This change ensures that the same mac-address is used. This is
achieved by allocating the mac-address based on the virtual-network the
port is attached to.

Change-Id: Ie96bd757f75f78b4aaa1c2a87a0e0ed2f902af01
Closes-Bug: #1460767