Schema transformer scaling improvements
With the help of profiler, Ajay and I found that most of the time is being spent
in walking all the VMIs for each network operations to find the VMIs on that
network and VMIs which correspond to a service instance. We added indexing on
VMIs for these two items and after that it finished significantly faster (~30
minutes).

Change-Id: I1d1e2110f12016c66b9db83ebd95c45d1a5e6e2d
Closes-Bug: 1460335
(cherry picked from commit 5cc9df6968d6ae5d25eb1ea6fc0eefe888e4d943)