Don't send XmppPeerInfo UVE on table subscribe/unsubscribe
The UVE has a list of tables to which the xmpp peer is registered.
Building and sending the entire list of tables each time the peer
subscribes and unsubscribes to a table is very expensive.

Do not send the UVE for the time being. Will resume sending the UVE
after new infra to make incremental updates to a list is available.

Change-Id: I1905755f665432e5cfc0229099e88bcc3bbf6b4c
Partial-Bug: 1456284