* Copy ARP mac, interface and resolved flag in Sync() API of route
Routes and nexthops dependent on ARP route are updated, when
any variable in ARP nexthop are changed. When ARP timer is
expired, nexthop gets marked as invalid and route Sync is triggered,
in this sync() API arp_mac and valid flags were not updated, hence
eventually when ARP resolution succeds, there would be no change in
agent path and the dependent nexthop would not be resolved.
Test case for same
Closes-bug:#1457337

Change-Id: I5c5c168367ce706a347a93f9f4852f8c5b68e95b