Fix valgrind reported issue in HTTP client
check_multi_info is invoked from "http client" and asio contexts.
Calling delete of connection info from here causes problems due to this.
Moving this such that if a new connection is created and if old
connection info is available, it is deleted. The last one is deleted when
HttpConnection is removed. Rechecked valgrind with this fix.

Change-Id: I0d2592b6ddb098d8f86e8036b0c1aa5c979627ac
closes-bug: 1461910