Do not club updates and deletes in a single message to ifmap
Because of a bug in irond, if update and delete of the same object received by it, it quashes it and the object is not deleted. Till that is fixed, we will split updates and deletes in separate messages.

Change-Id: I5504477351b705837679242554ad2d7ffd529f98
Closes-Bug: 1461058
Closes-Bug: 1461626
(cherry picked from commit 3086f689cfe7135897ce951c95732c353012b0fd)