Implement heuristic to disable SchedulingGroup splitting
It's not worth trying to split a group with a very large number of
(ribout, peer) members.  The chances that such a group can be split
are pretty low and the amount of effort spent in figuring it out is
quite large.

Implement a heuristic/hack to make a group split ineligible once it
has more than a certain number of members. This is a sticky property
i.e. the group can never be split once this property gets set.

If/when more parallelism is required on the xmpp send side, it can
be achieved by creating multiple RibOuts for the same tables based
on a hash of the agent name/address.

Change-Id: I601f9aadc4e21a835b7e320f8aae430b5589929d
Partial-Bug: 1461322