The NodeTableExport() should call ConfigChanged() for a node-delete.
If the node gets revived (add followed by delete followed by add), we could have
cases where the CRC in the second add is the same as CRC in the first add,
causing 'bool change' to be false. But, the node has really changed since there
was a delete in the middle.

Closes-Bug: 1462549

Change-Id: Ife1f98ef8bc35826b829d2d122068edcf092469e