Fix concurrency issue in XmppSession::WriteReady
It's possible that XmppSession::WriteReady gets called after the
xmpp connection for the session has already been deleted. Prevent
WriteReady from from accessing freed memory by clearing the back
pointer to the xmpp connection in the session when the session in
the xmpp connection is cleared.

The above fix is not sufficient since there's a concurrency issue
wherein XmppSession::WriteReady could try to access the connection
the connection is trying to clear the back pointer in the session.

Fix is to make XmppSession::WriteReady enqueue sessions to a work
queue and process the work queue in context of bgp::Config task.
Add XmppConnectionManager class so that this can be done for both
XmppClient and XmppServer. Other common functionality can be moved
into this class in future.

Remove dead code for XmppStream.

Change-Id: Id9ea3311e6618a0afc20058c3571330a1627fcec
Closes-Bug: 1463122