Allow zookeeper heartbeat to expire if determined by the application
At init time, api server may be busy doing resync. At this time, it is ok if
zookeeper heartbeat expires. Added callback in zkclient to be called when kazoo
state goes to lost. API server sets it to a stub function initially, then resets
it after init is done.

Change-Id: Icd91cfb9637a8b0086a5486722a47e26a875acbb
Partial-Bug: 1463270