Fix race condition in SchedulingGroup::Worker
The WorkQueue in the SchedulingGroup contains naked pointers to RibOuts
and IPeerUpdates without any reference counting.  If the IPeerUpdate or
RibOut is removed from the SchedulingGroup while there are corresponding
WorkPeer or WorkRibOut entries in the WorkQueue, the Worker::Run method
crashes when it processes the WorkPeer or WorkRibOut.

Fix is to invalidate the relevant WorkQueue entries when an IPeerUpdate
or RibOut is removed from the SchedulingGroup.

Change-Id: I496fc91b85c3c55786676639c0a146cbda2d730c
Closes-Bug: 1464106