l3 vrf and l2 evpn vrf config both need to be configured
Currently, DM configures only one vrf for a a virtual network either l2 evpn or
l3 vrf config but not both depending on the vxlan vni identifier.

It must configure both for a private network.
For a public network, only l3 vrf config is required.
import-export statment is not required for l2 vrf.

Fixed the issue - self paramter is minning in get_commit_stats()

Closes-Bug: #1464446
Closes-Bug: #1464786

Change-Id: I6eb9763342631ba6a9235cc4145ea180604e5c8a