il3 vrf and l2 evpn vrf config both need to be configured
Currently, DM configures only one vrf for a a virtual network either l2 evpn or
l3 vrf config but not both depending on the vxlan vni identifier.

It must configure both for a private network.
For a public network, only l3 vrf config is required.
import-export statment is not required for l2 vrf.

Change-Id: I7821aa535622df9ce7ee780fd6a44abc680aaec8
Closes-Bug: #1464446