Additional verification for WorkRibOut/WorkPeer invalidate logic
The new tests identified an issue where WorkRibOuts/WorkPeers get
invalidated when splitting a SchedulingGroup.

Fix this by making sure that the invalidation logic kicks in only
when RibOut or IPeerUpdate is removed from SchedulingGroupManager,
not when it's removed from a SchedulingGroup since the latter may
happen because of a split/merge.

Change-Id: I9fde07df65b6dec3c1cb58376c761350b64d0ae2
Closes-Bug: 1464106