Allow to start the vnc_openstack in a separate process
In order to distribute the cpu load of the vnc api
vnc openstack, this patch allows to start the vnc_openstack
in a separate process.
This behavior is optional and can be set by setting the
new parameter vnc_openstack_port. The vnc_api still
continues to work as usual providing the both APIs.

Closes-Bug: #1465325
Change-Id: Ifad441a930512f82f550fce274de473062ea0f5e