Closes-Bug: #1466363 Use fork-exec and waitpid() instead of platform specific system() call
While pausing bgp_stress_test, a python shell is forked. If system() call is used,
then SIGCHLD is not caught if the python process is killed. That causes the forked
python process to go to defunct state and bgp_stress_test to hang.

Found during becaa's scale test scripts run

Change-Id: I7d52af0c41b133c70938ebc65d62858ab299e056