Reset the stitched MAC if no replacement route found
When an INET/INET6 route is deleted, agent gives flags and stitched-mac
from the replacmenet route.

When there is no replacement route, agent resets all the flags but does not
change the stitched mac. In this case, vrouter will return error since
stitched-mac is not valid for the replacement route.

Fix:
Ensure the stitched mac is reset when replacement route is not found

Change-Id: Ia669baf7cd895f97fb8de290db94253f7887a6a1
Closes-Bug: #1466596