* Dont apply interface VRF assign rule when packet gets NATed
Interface VRF assign rules are ysed to steer traffic thru service
instances in service chain. If a packet gets NATed then destination
VRF should be set to native VRF of interface to which the floating
ip belong, and in this case there is no need to apply VRF translation
rule. Test case to verify the same.
Closes-bug:#1465728

Change-Id: I19ae3dda079269f7f7d041ebfd110c37c01f16c1