[SVC monitor] Add mechanism to prevent flapping VM SI scheduling
For any reason, a vrouter status can flap between functional and
non-fonctional that can generate unnecessary re-scheduling of
all SI VM.

That fix adds a parameter to permit to decide how many times a vrouter
needs to be non-functional before removing and re-schedule its SI VM
(named 'retry_before_scheduling' in SCHEDULER section and default to
1 time).

It also permits to set the health check interval in parameter (named
'cleanup_delay' in DEFAULT section and default to 5 seconds).

That patch also correct a bug when a VM is remove from a vrouter
and unit tests which are broken (how that's possible ??).

Closes-Bug: #1466588
Change-Id: I2c6b591de264ddd211e6c1daea6b48bf827f7540